+++ crates/factor-wasi/Cargo.toml
@@ -10,6 +10,7 @@ bytes = { workspace = true }
 spin-common = { path = "../common" }
 spin-factors = { path = "../factors" }
 tokio = { workspace = true, features = ["sync"] }
+tracing = { workspace = true }
 wasmtime = { workspace = true }
 wasmtime-wasi = { workspace = true }
